+++ packages/client-utils/src/index.ts
@@ -5,9 +5,6 @@ export { mapApiTransaction } from './mappers/api-transaction-mapper.js';
 export { mapKeyringTransaction } from './mappers/keyring-transaction-mapper.js';
 export { mapLocalTransaction } from './mappers/local-transaction-mapper.js';
 export { mapRampsOrder } from './mappers/ramps-order-mapper.js';
-export type {
-  RampOrderPaymentDetail,
-  RampsOrderLike,
-} from './mappers/ramps-order-mapper.js';
+export type { RampsOrderLike } from './mappers/ramps-order-mapper.js';
 
 export type * from './types.js';
