+++ packages/client-utils/src/mappers/ramps-order-mapper.ts
@@ -1,18 +1,15 @@
 import type { CaipChainId } from '@metamask/utils';
 import { isCaipChainId } from '@metamask/utils';
 
-import type { ActivityItem, Fee, FiatAmount, Status, TokenAmount } from '../types.js';
+import type {
+  ActivityItem,
+  Fee,
+  FiatAmount,
+  RampOrderPaymentDetail,
+  Status,
+  TokenAmount,
+} from '../types.js';
 import { formatChainIdToCaip } from './helpers/caip.js';
-
-/**
- * Bank transfer instruction fields attached to a ramp order by providers
- * that require manual payment (e.g. SEPA, wire transfer).
- */
-export type RampOrderPaymentDetail = {
-  fiatCurrency: string;
-  paymentMethod: string;
-  fields: { name: string; id: string; value: string }[];
-};
 
 /**
  * Extracts the CAIP-2 chain id from a CAIP-19 asset id
